pyproject.toml
[tool.poetry]
name = "trees-everywhere"
version = "0.1.0"
description = "I see trees, trees everywhere."
authors = ["leonardo-kazu <lkazu@duck.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.2"
psycopg = {extras = ["binary"], version = "^3.1.18"}


[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^24.2.0"
pre-commit = "^3.6.2"
commitizen = "^3.15.0"
python-dotenv = "^1.0.1"
factory-boy = "^3.3.0"
faker = "^23.2.1"




[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "poetry"
update_changelog_on_bump = true
major_version_zero = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
